FORMS-720: fix control

+++ src/java/fr/paris/lutece/plugins/forms/validation/AbstractDateValidator.java
@@ -72,7 +72,7 @@ public abstract class AbstractDateValidator extends AbstractValidator
             Date dateControl = new Date( dateFormat.parse( control.getValue( ) ).getTime( ) );
             for ( Response response : questionResponse.getEntryResponse( ) )
             {
-                Date date = new Date( Integer.parseInt( response.getResponseValue( ) ) );
+                Date date = new Date( Long.parseLong( response.getResponseValue( ) ) );
                 if ( !validateDate( dateControl, date ) )
                 {
                     return false;
